requirement.txt
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
boto3==1.34.74
botocore==1.34.74
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
dataclasses-json==0.5.14
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.8.0
filelock==3.13.3
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
importlib-metadata==6.11.0
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.1
langchain==0.0.240
langchain-community==0.0.29
langchain-core==0.1.36
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.0.92
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
nest-asyncio==1.5.8
networkx==3.2.1
numexpr==2.9.0
numpy==1.26.4
openai==1.14.3
openapi-schema-pydantic==1.2.4
orjson==3.10.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==4.21.12
psutil==5.9.7
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pydantic==1.10.14
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
Pympler==1.0.1
pypdf==4.1.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
s3transfer==0.10.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.6.1
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.23
stack-data==0.6.3
streamlit==1.24.0
streamlit-chat==0.1.1
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.4.0.dev20240606
torchvision==0.19.0.dev20240606
tornado==6.4
tqdm==4.66.2
traitlets==5.14.0
transformers==4.39.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==4.3.1
urllib3==1.26.18
validators==0.24.0
wcwidth==0.2.12
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
